Added the facets volume by term, volume by mentioned countries, volume by funders, with:
- bar chart of nodes by volume of items
- bar chart of links by volume of items
- network panel:
  - force layout of nodes
  - histogram of nodes degree
  - histogram of nodes volume
  - histogram of links volume

We show just the 100 nodes and 100 links for performance reasons (later on this will probably be exposed to the user via histogram interactions – selecting/brushing bins).

We get the nodes using an accessor, so that we can control when to do if the node can't be found (for example by providing a default value via the accessor).

Clicking on a node on the nodes bar chart or on the force can do 2 possible things depending on the switch on the top-right:
  - `exclude`: filtering out items with that term in `terms`
  - `ignore`: hide the node from the views without actually filtering out items with that term. This is technically not a `selection` but a `mask`, we'll split selections and masks in a new issue

We use the same color scale to represent the volume for nodes and links in barcharts, force and histograms.

Histograms have now a vertical orientation ('bottom-up' – default, vs 'top-down') and a horizontal orientation ('left-right' – default, or 'right-left') (#235)

Updated the Histogram scale so that if the order of magnitude of max - min value is greater than 2 we use logarithmic scales (it'd be basically impossible to see the legend colours if we have items with very different O()s. (#236)

Barchart: added an `interactive` prop to distinguish clickable items and hoverable ones (#237)

Dev details: polymorphic `makeIsWithin`, works for range and bounds
Fix the map not resetting when sending an undefined value

Closes #1
Closes #232
Closes #233
Closes #235
Closes #236
Closes #237